map-spat-broadcaster, all transceiver-components are set to read the configuration file from /nojournal/bin, the configuration file renamed mmitss-phase3-master-config.json, configuration file errors fixed, priority-request-server arm static_int issue fixed. Now need to remove hardcoded ip addresses from all remaining files before merging to develop

+++ src/rsu/priority-request-server/M_PriorityRequestServer.cpp
@@ -294,7 +294,7 @@ void sendSSM(LinkedList<ReqEntry> ReqList, const int IntersectionID, UdpSocket M
 
         for (int i = 0; !ReqList.EndOfList(); i++)
         {
-            jsonObject["SignalStatus"]["requestorInfo"][i]["vehicleID"] = ReqList.Data().VehID;
+            jsonObject["SignalStatus"]["requestorInfo"][i]["vehicleID"] = static_cast<int>ReqList.Data().VehID;
             jsonObject["SignalStatus"]["requestorInfo"][i]["requestID"] = 0;
             jsonObject["SignalStatus"]["requestorInfo"][i]["msgCount"] = ReqList.Data().iMsgCnt;
             jsonObject["SignalStatus"]["requestorInfo"][i]["basicVehicleRole"] = ReqList.Data().VehClass;
